Test for the multiple molecule loader
