+++ test/poltype/output/ethanol_lf_poltype.txt
@@ -1,0 +1,24 @@
+1 9 2
+2 3 1
+3 2 1
+4 2 5
+5 2 4
+6 3 2
+7 3 2
+8 3 2
+9 1 2
+
+A
+1 0.812
+2 1.620
+3 1.415
+4 0.480
+5 0.480
+6 0.480
+7 0.480
+8 0.480
+9 0.457
+
+2
+N
+Y
