+++ test/poltype/output/ethanol_lf_sulley.txt
@@ -1,0 +1,9 @@
+1 9 2
+2 3 1
+3 2 1
+4 2 5
+5 2 4
+6 3 2
+7 3 2
+8 3 2
+9 1 2
