+++ test/poltype/structures/aspirin_ethanol.xyz
@@ -1,0 +1,33 @@
+    30 
+     1   O   1.2333    0.5540    0.7792    401     5    12
+     2   O  -0.6952   -2.7148   -0.7502    401    11    21
+     3   O   0.7958   -2.1843    0.8685    401    11
+     4   O   1.7813    0.8105   -1.4821    401    12
+     5   C  -0.0857    0.6088    0.4403    401     6     7     1
+     6   C  -0.7927   -0.5515    0.1244    401     8    11     5
+     7   C  -0.7288    1.8464    0.4133    401     9    14     5
+     8   C  -2.1426   -0.4741   -0.2184    401    10    15     6
+     9   C  -2.0787    1.9238    0.0706    401    10    16     7
+    10   C  -2.7855    0.7636   -0.2453    401    17     8     9
+    11   C  -0.1409   -1.8536    0.1477    401     2     3     6
+    12   C   2.1094    0.6715   -0.3113    401    13     1     4
+    13   C   3.5305    0.5996    0.1635    401    18    19    20    12
+    14   H  -0.1851    2.7545    0.6593    401     7
+    15   H  -2.7247   -1.3605   -0.4564    401     8
+    16   H  -2.5797    2.8872    0.0506    401     9
+    17   H  -3.8374    0.8238   -0.5090    401    10
+    18   H   3.7290    1.4184    0.8593    401    13
+    19   H   4.2045    0.6969   -0.6924    401    13
+    20   H   3.7105   -0.3659    0.6426    401    13
+    21   H  -0.2555   -3.5916   -0.7337    401     2
+
+
+    31   O  -1.1712    0.2997    0.0000    401    32    39  
+    32   C  -0.0463   -0.5665    0.0000    401    33    34    35    31
+    33   C   1.2175    0.2668    0.0000    401    36    37    38    32 
+    34   H  -0.0958   -1.2120    0.8819    401    32
+    35   H  -0.0952   -1.1938   -0.8946    401    32
+    36   H   2.1050   -0.3720   -0.0177    401    33
+    37   H   1.2426    0.9307   -0.8704    401    33
+    38   H   1.2616    0.9052    0.8886    401    33
+    39   H  -1.1291    0.8364    0.8099    401    31
