+++ test/poltype/structures/ethanol.sdf
@@ -1,0 +1,110 @@
+702
+  -OEChem-10182408233D
+
+  9  8  0     0  0  0  0  0  0999 V2000
+   -1.1712    0.2997    0.0000 O   0  0  0  0  0  0  0  0  0  0  0  0
+   -0.0463   -0.5665    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0
+    1.2175    0.2668    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0
+   -0.0958   -1.2120    0.8819 H   0  0  0  0  0  0  0  0  0  0  0  0
+   -0.0952   -1.1938   -0.8946 H   0  0  0  0  0  0  0  0  0  0  0  0
+    2.1050   -0.3720   -0.0177 H   0  0  0  0  0  0  0  0  0  0  0  0
+    1.2426    0.9307   -0.8704 H   0  0  0  0  0  0  0  0  0  0  0  0
+    1.2616    0.9052    0.8886 H   0  0  0  0  0  0  0  0  0  0  0  0
+   -1.1291    0.8364    0.8099 H   0  0  0  0  0  0  0  0  0  0  0  0
+  1  2  1  0  0  0  0
+  1  9  1  0  0  0  0
+  2  3  1  0  0  0  0
+  2  4  1  0  0  0  0
+  2  5  1  0  0  0  0
+  3  6  1  0  0  0  0
+  3  7  1  0  0  0  0
+  3  8  1  0  0  0  0
+M  END
+> <PUBCHEM_COMPOUND_CID>
+702
+
+> <PUBCHEM_CONFORMER_RMSD>
+0.4
+
+> <PUBCHEM_CONFORMER_DIVERSEORDER>
+1
+
+> <PUBCHEM_MMFF94_PARTIAL_CHARGES>
+3
+1 -0.68
+2 0.28
+9 0.4
+
+> <PUBCHEM_EFFECTIVE_ROTOR_COUNT>
+0
+
+> <PUBCHEM_PHARMACOPHORE_FEATURES>
+2
+1 1 acceptor
+1 1 donor
+
+> <PUBCHEM_HEAVY_ATOM_COUNT>
+3
+
+> <PUBCHEM_ATOM_DEF_STEREO_COUNT>
+0
+
+> <PUBCHEM_ATOM_UDEF_STEREO_COUNT>
+0
+
+> <PUBCHEM_BOND_DEF_STEREO_COUNT>
+0
+
+> <PUBCHEM_BOND_UDEF_STEREO_COUNT>
+0
+
+> <PUBCHEM_ISOTOPIC_ATOM_COUNT>
+0
+
+> <PUBCHEM_COMPONENT_COUNT>
+1
+
+> <PUBCHEM_CACTVS_TAUTO_COUNT>
+1
+
+> <PUBCHEM_CONFORMER_ID>
+000002BE00000001
+
+> <PUBCHEM_MMFF94_ENERGY>
+-1.3096
+
+> <PUBCHEM_FEATURE_SELFOVERLAP>
+10.148
+
+> <PUBCHEM_SHAPE_FINGERPRINT>
+139733 1 18410856563987595332
+21015797 1 18192151726869919494
+
+> <PUBCHEM_SHAPE_MULTIPOLES>
+55.87
+1.48
+0.75
+0.58
+0.08
+0.07
+0
+-0.26
+0
+-0.01
+0
+0.04
+0.01
+0
+
+> <PUBCHEM_SHAPE_SELFOVERLAP>
+84.621
+
+> <PUBCHEM_SHAPE_VOLUME>
+41.5
+
+> <PUBCHEM_COORDINATE_TYPE>
+2
+5
+10
+
+$$$$
