+++ test/poltype/test_with_poltype.py
@@ -147,36 +147,6 @@ def test_aspirin():
         test_path + "output/aspirin_lf_poltype.txt"
     )
 
-def test_aspirin_xyz():
-
-    test_path = os.getcwd() + "/test/poltype/"
-    
-    # Poltype
-    mol = open_sdf_convert_to_mol(
-        test_path + "structures/aspirin.sdf",
-        test_path + "structures/aspirin.mol"
-    )
-    poltype = Poltype(
-        mol,
-        peditinfile=test_path + "output/aspirin_lf_poltype.txt",
-        molstructfname=test_path + "structures/aspirin.mol",
-        paramfile=test_path + "polarize.prm"
-    )
-
-    gen_peditinfile(poltype, mol)
-    
-    # Sulley
-    mol = load_molecule_from_tinker_xyz(test_path + "structures/aspirin.xyz")
-    generate_local_frame(
-        mol=mol,
-        filename=test_path + "output/aspirin_lf_sulley.txt"
-    )
-
-    assert is_same_local_frame(
-        test_path + "output/aspirin_lf_sulley.txt",
-        test_path + "output/aspirin_lf_poltype.txt"
-    )
-
 
 def open_sdf_convert_to_mol(sdf_file, mol_file):
     obConversion = openbabel.OBConversion()
